Pipelines: lint and bypass (#12)
When status checks are required but not executed (e.g. due to `paths`), then PRs can't be merged. In order for that a second workflow needs to be defined that has the same name and job name that always succeeds, in case the "real" workflow isn't triggered. See here for more details: https://docs.github.com/en/repositories/configuring-branches-and-merges-in-your-repository/defining-the-mergeability-of-pull-requests/troubleshooting-required-status-checks#handling-skipped-but-required-checks

Issue: #6